@lemire 2020-10-16
@jacksonrnewhouse The change is very reasonable, and I believe your positive results.

I would recommend merging this, but I would suggest two fixes:

1. We are still supporting Go 1.12, and your PR breaks on 1.12. I am game to drop 1.12, but it seems a tad early to do so. So either argue for us dropping 1.12, or please make your PR work under 1.12.

2. Can you add maybe only one line of comment so that a reader can quickly understand the rationale of your new branch? It can be just one short sentence. (It is to ease maintenance.)