@lemire 2020-10-16
@jacksonrnewhouse Let me know if you recommend merging.