@lemire 2020-10-19
Looks like high quality code, unlikely to introduce new bugs. Likely to improve performance in some cases. Merging.